+++ spec/rubocop/cop/rails/unknown_env_spec.rb
@@ -14,19 +14,32 @@ RSpec.describe RuboCop::Cop::Rails::UnknownEnv, :config do
   end
 
   it 'registers an offense for typo of environment name' do
-    expect_offense(<<-RUBY)
+    expect_offense(<<~RUBY)
       Rails.env.proudction?
-                ^^^^^^^^^^^ Unknown environment `proudction?`. Did you mean `production?`?
+                ^^^^^^^^^^^ Unknown environment `proudction`. Did you mean `production`?
       Rails.env.developpment?
-                ^^^^^^^^^^^^^ Unknown environment `developpment?`. Did you mean `development?`?
+                ^^^^^^^^^^^^^ Unknown environment `developpment`. Did you mean `development`?
       Rails.env.something?
-                ^^^^^^^^^^ Unknown environment `something?`.
+                ^^^^^^^^^^ Unknown environment `something`.
+    RUBY
+  end
+
+  it 'registers an offense for typo of environment name with `==` operator' do
+    expect_offense(<<~RUBY)
+      Rails.env == 'proudction'
+                   ^^^^^^^^^^^^ Unknown environment `proudction`. Did you mean `production`?
+      'developpment' == Rails.env
+      ^^^^^^^^^^^^^^ Unknown environment `developpment`. Did you mean `development`?
+
+      'something' === Rails.env
+      ^^^^^^^^^^^ Unknown environment `something`.
     RUBY
   end
 
   it 'accepts correct environment name' do
-    expect_no_offenses(<<-RUBY)
+    expect_no_offenses(<<~RUBY)
       Rails.env.production?
+      Rails.env == 'production'
     RUBY
   end
 end